Add subcommands dealing with the model repository
The model repository is a central github repository containing a variety
of freely licensed recognition models for kraken and clstm. These are
verified to work for certain types of input and are ensured to run with
the current version of kraken/clstm. By publication they are accessible
from any kraken installation.