Added the canvas method for drawing rectangles over the image
Added the canvas for drawing rectangles over the image in a more easy way

however it does not seem to work so well. I need to work on it a little more
Name space of the ids are also changed.
Most underscores are changed to dashes to keep up with the general style of
html5