better treatment of training/validation set codecs
Current behavior is that validation set only labels are added to the
training set. This would produce untrained output labels in the network.
The new behavior is split codecs for training and validation with the
labels that aren't situated at the end of the validation codec. This
allows decoding during the validation run without polluting the training
codec.