Search segmentation model in XDG_BASE_DIR
According to the documentation, models are placed in `$XDG_BASE_DIR`
and can be accessed using their name.

While this worked for OCR models, segmentation models were not searched
in the kraken config folder. This could be confusing for users.

Now segmentation models can be specified without a path and will be
found in the config folder.

Update also the comment for the OCR models.

Signed-off-by: Stefan Weil <sw@weilnetz.de>